% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perfomances.R
\name{conindex}
\alias{conindex}
\title{\code{cindex}}
\usage{
conindex(obj, Y)
}
\arguments{
\item{obj}{[\code{survivalsvmprediction}]\cr
Object of class \code{survivalsvmprediction}.}

\item{Y}{[\code{vector}(1)]\cr
A \code{Surv()} object with true survival times observed.}
}
\value{
[\code{Integer}]
Concordance index.
}
\description{
computes the concordance index.
}
\keyword{internal}
