% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diffmatrix.R
\name{Diffmatrix}
\alias{Diffmatrix}
\title{\code{Diffmatrix}.}
\usage{
Diffmatrix(Type = NULL, Mat = NULL)
}
\arguments{
\item{Type}{[\code{character(1)}]\cr
Indicates which difference is performed. This must be one of \code{\link{makediff1}}, \code{\link{makediff2}} other \code{\link{makediff3}}.}

\item{Mat}{[\code{matrix(1)}]\cr
Matrix used to perfom differences.}
}
\value{
[\code{Diffmatrix(1)}]
Mutated object of class \code{Diffmatrix} containing elements:\cr
 \tabular{ll}{
   \code{Type} \tab type of differences bildet between neighbors and \cr
   \code{Mat} \tab matrix used to perform differences between comparable data points. \cr
 }
}
\description{
constructs objects of class \code{Diffmatrix}.
}
\keyword{internal}

