% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vanbelle1.R
\name{getAlpha.default}
\alias{getAlpha.default}
\title{\code{VB1FitObj} (ranking approach)}
\usage{
\method{getAlpha}{default}(vb1o)
}
\arguments{
\item{vb1o}{[\code{VB1FitObj}]\cr
Object taken in the argument.}
}
\value{
\code{NULL}.
}
\description{
Accessor for the field \code{Alpha} for the object taken in an argument.
}
\author{
Cesaire J. K. Fouodo
}
\keyword{internal}

