% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vanbelle1.R
\name{getAlpha.VB1FitObj}
\alias{getAlpha.VB1FitObj}
\title{\code{VB1FitObj} (ranking approach)}
\usage{
\method{getAlpha}{VB1FitObj}(vb1o)
}
\arguments{
\item{vb1o}{[\code{VB1FitObj}]\cr
Object taken in the argument.}
}
\value{
[\code{vector(1)}]
Alpha field of the object of class \code{VB1FitObj} taken in the argument.
}
\description{
Creator of the generic accessor \code{Beta}.
}
\author{
Cesaire J. K. Fouodo
}
\keyword{internal}

