% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{getb0}
\alias{getb0}
\title{\code{RegFitObj} (regression approach)}
\usage{
getb0(rfo)
}
\arguments{
\item{rfo}{[\code{RegFitObj}(1)]\cr
object taken in the argument.}
}
\description{
Creator of the generic accessor \code{getb0}.
}
\author{
Cesaire J. K. Fouodo
}
\keyword{internal}

