% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{getb0.RegFitObj}
\alias{getb0.RegFitObj}
\title{Class \code{RegFitObj} (regression approach)}
\usage{
\method{getb0}{RegFitObj}(rfo)
}
\arguments{
\item{rfo}{[\code{RegFitObj}(1)]\cr
object taken in the argument.}
}
\value{
[\code{numeric(1)}]
the offset
}
\description{
Creator of the generic accessor \code{b0}.
}
\author{
Cesaire J. K. Fouodo
}
\keyword{internal}
