% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{getBeta.RegFitObj}
\alias{getBeta.RegFitObj}
\title{\code{RegFitObj}}
\usage{
\method{getBeta}{RegFitObj}(rfo)
}
\arguments{
\item{rfo}{[\code{RegFitObj}(1)]\cr
object taken in the argument.}
}
\value{
[\code{vector(1)}]
Beta field of the object of class \code{RegFitObj} taken in the argument.
}
\description{
Creator of the generic accessor \code{Beta}.
}
\author{
Cesaire J. K. Fouodo
}
\keyword{internal}
