% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid.R
\name{getBetastar.default}
\alias{getBetastar.default}
\title{\code{Hybrid} (hybrid approach)}
\usage{
\method{getBetastar}{default}(hybo)
}
\arguments{
\item{hybo}{[\code{Hybrid(1)}]\cr
Object taken in the argument.}
}
\value{
\code{NULL}.
}
\description{
Accessor for the field \code{Betastar} for the object taken in an argument.
}
\author{
Cesaire J. K. Fouodo
}
\keyword{internal}
