% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid.R
\name{getBetastar.HybridObj}
\alias{getBetastar.HybridObj}
\title{\code{Hybrid} (hybrid approach)}
\usage{
\method{getBetastar}{HybridObj}(hybo)
}
\arguments{
\item{hybo}{[\code{Hybrid(1)}]\cr
Object taken in the argument.}
}
\value{
[\code{vector(1)}]
Betastar field of the object of class \code{Hybrid} taken in the argument.
}
\description{
Creator of the generic accessor \code{Betastar}.
}
\author{
Cesaire J. K. Fouodo
}
\keyword{internal}

