% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{getBinca.default}
\alias{getBinca.default}
\title{\code{Kernel}}
\usage{
getBinca.default(kern)
}
\arguments{
\item{kern}{[\code{\link{Kernel}(1)}]\cr
Object taken in the argument.}
}
\value{
\code{NULL}.
}
\description{
Accessor for the field \code{Bincat} for the object taken in the argument.
}
\author{
Cesaire J. K. Fouodo
}
\keyword{internal}
