% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{getBincat.Kernel}
\alias{getBincat.Kernel}
\title{\code{Kernel}}
\usage{
\method{getBincat}{Kernel}(kern)
}
\arguments{
\item{kern}{[\code{\link{Kernel}(1)}]\cr
Object of class \code{Kernel}.}
}
\value{
Index of binary/categorical variables.
}
\description{
Accessor for the field \code{Bincat} of the object of class \code{Kernel} taken in an argument.
}
\author{
Cesaire J. K. Fouodo
}
\keyword{internal}

