% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{getBincat}
\alias{getBincat}
\title{\code{Kernel}}
\usage{
getBincat(kern)
}
\arguments{
\item{kern}{[\code{\link{Kernel}(1)}]\cr
Object taken in the argument.}
}
\description{
Creator of the generic accessor \code{getBincat}.
}
\author{
Cesaire J. K. Fouodo
}
\keyword{internal}
