% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid.R
\name{getDelta.HybridObj}
\alias{getDelta.HybridObj}
\title{\code{Hybrid} (hybrid approach)}
\usage{
\method{getDelta}{HybridObj}(hybo)
}
\arguments{
\item{hybo}{[\code{Hybrid(1)}]\cr
Object taken in an argument.}
}
\value{
[\code{vector(1)}]
Delta field of the object of class \code{Hybrid} taken in the argument.
}
\description{
Creator of the generic accessor \code{Delta}.
}
\author{
Cesaire J. K. Fouodo
}
\keyword{internal}
