% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid.R
\name{getDelta}
\alias{getDelta}
\title{\code{Hybrid} (hybrid approach)}
\usage{
getDelta(hybo)
}
\arguments{
\item{hybo}{[\code{Hybrid(1)}]\cr
Object taken in the argument.}
}
\description{
Creator of the generic accessor \code{getDelta}.
}
\author{
Cesaire J. K. Fouodo
}
\keyword{internal}
