% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vanbelle1.R
\name{getDifMat.VB1FitObj}
\alias{getDifMat.VB1FitObj}
\title{\code{VB1FitObj} (ranking approach)}
\usage{
\method{getDifMat}{VB1FitObj}(vb1o)
}
\arguments{
\item{vb1o}{[\code{VB1FitObj}]\cr
object taken in argument.}
}
\value{
DifMat [\code{\link{Diffmatrix}(1)}]
field of the object of class \code{VB1FitObj} taken in the argument.
}
\description{
access to the matrix in the \code{DifMat} field.
}
\author{
Cesaire J. K. Fouodo
}
\keyword{internal}

