% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{getKernel}
\alias{getKernel}
\title{\code{RegFitObj} (regression approach)}
\usage{
getKernel(rfo)
}
\arguments{
\item{rfo}{[\code{RegFitObj}(1)]\cr
object taken in the argument.}
}
\description{
Creator of the generic accessor \code{getKernel}.
}
\author{
Cesaire J. K. Fouodo
}
\keyword{internal}
