% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{getKernel.RegFitObj}
\alias{getKernel.RegFitObj}
\title{\code{RegFitObj}}
\usage{
\method{getKernel}{RegFitObj}(rfo)
}
\arguments{
\item{rfo}{[\code{RegFitObj}(1)]\cr
object taken in the argument.}
}
\value{
[\code{Kernel}]
kernel.
}
\description{
Creator of the generic accessor \code{Kernel}.
}
\author{
Cesaire J. K. Fouodo
}
\keyword{internal}

