% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vanbelle1.R
\name{getKernel.VB1FitObj}
\alias{getKernel.VB1FitObj}
\title{\code{VB1FitObj} (ranking approach)}
\usage{
\method{getKernel}{VB1FitObj}(vb1o)
}
\arguments{
\item{vb1o}{[\code{VB1FitObj}]\cr
Object taken in argument.}
}
\value{
Kernel [\code{\link{Kernel}(1)}]
Field of the object of class \code{VB1FitObj} taken in the argument.
}
\description{
access the object of class \code{Kernel} in the \code{Kernel} field.
}
\author{
Cesaire J. K. Fouodo
}
\keyword{internal}

