% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{getKernpar.default}
\alias{getKernpar.default}
\title{\code{Kernel}}
\usage{
\method{getKernpar}{default}(kern)
}
\arguments{
\item{kern}{[\code{\link{Kernel}(1)}]\cr
Object taken in the argument.}
}
\value{
\code{NULL}.
}
\description{
Accessor for the field \code{Kernpar} for the object taken in an argument.
}
\author{
Cesaire J. K. Fouodo
}
\keyword{internal}
