% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{getKernpar.Kernel}
\alias{getKernpar.Kernel}
\title{\code{Kernel}}
\usage{
\method{getKernpar}{Kernel}(kern)
}
\arguments{
\item{kern}{[\code{\link{Kernel}(1)}]\cr
Object of class \code{Kernel}.}
}
\value{
[\code{vector(1)}]
The kernel parameters.
}
\description{
Accessor for the field \code{Kernpar} of the object of class \code{Kernel} taken in an argument.
}
\author{
Cesaire J. K. Fouodo
}
\keyword{internal}

