% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perfomances.R
\name{getLogrank}
\alias{getLogrank}
\title{\code{getLogrank}}
\usage{
getLogrank(obj, t, delta)
}
\arguments{
\item{obj}{[\code{survivalsvmprediction}(1)]\cr
Object of class \code{survivalsvmprediction}.}

\item{t}{[\code{numeric}(1)]\cr
Numeric vector (of survival times).}

\item{delta}{[\code{vector}(1)]\cr
Binary vector (of status).}
}
\value{
list of:
\tabular{ll}{
 \code{chi_sq} \tab chi-squared statistic and \cr
 \code{chi_p} \tab chi-squared probality. \cr
}
}
\description{
computes the logrank statistic.
}

