% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diffmatrix.R
\name{getMat.Diffmatrix}
\alias{getMat.Diffmatrix}
\title{\code{Diffmatrix}.}
\usage{
\method{getMat}{Diffmatrix}(obj)
}
\arguments{
\item{obj}{[\code{\link{Diffmatrix}(1)}]\cr
Object of class \code{Diffmatrix}.}
}
\value{
The matrix used to perform differences between comparable data points.
}
\description{
To get the matrix used to perform differences between comparable data points for object of class Diffmatrix.
}
\keyword{internal}
