% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{getMat.Kernel}
\alias{getMat.Kernel}
\title{\code{Kernel}}
\usage{
\method{getMat}{Kernel}(kern)
}
\arguments{
\item{kern}{[\code{\link{Kernel}(1)}]\cr
Object of class \code{Kernel}.}
}
\value{
[\code{matrix(1)}]
The kernel matrix.
}
\description{
Accessor for the field \code{Mat} of the object of class \code{Kernel} taken in an argument.
}
\author{
Cesaire J. K. Fouodo
}
\keyword{internal}

