% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{getMat}
\alias{getMat}
\title{\code{Kernel}}
\usage{
getMat(obj)
}
\arguments{
\item{obj}{[\code{Object(1)}]\cr
Object taken in the argument.}
}
\description{
Creator of the generic accessor \code{getMat}.
}
\author{
Cesaire J. K. Fouodo
}
\keyword{internal}
