% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{getOptMeth.default}
\alias{getOptMeth.default}
\title{\code{RegFitObj}}
\usage{
\method{getOptMeth}{default}(rfo)
}
\arguments{
\item{rfo}{[\code{RegFitObj}(1)]\cr
object taken in the argument.}
}
\value{
\code{NULL}.
}
\description{
Accessor for the field \code{OptMeth} for the object taken in an argument.
}
\author{
Cesaire J. K. Fouodo
}
\keyword{internal}

