% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{getOptMeth.RegFitObj}
\alias{getOptMeth.RegFitObj}
\title{Class \code{RegFitObj} (regression approach)}
\usage{
\method{getOptMeth}{RegFitObj}(rfo)
}
\arguments{
\item{rfo}{[\code{RegFitObj}(1)]\cr
object taken in the argument.}
}
\value{
[character(1)]
the named of the optimization program.
}
\description{
Creator of the generic accessor \code{OptMeth}.
}
\author{
Cesaire J. K. Fouodo
}
\keyword{internal}
