% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vanbelle1.R
\name{getOptMeth.VB1FitObj}
\alias{getOptMeth.VB1FitObj}
\title{\code{VB1FitObj} (ranking approach)}
\usage{
\method{getOptMeth}{VB1FitObj}(vb1o)
}
\arguments{
\item{vb1o}{[\code{VB1FitObj}]\cr
Object taken in argument.}
}
\value{
[\code{character(1)}]
OptMeth field of the object of class \code{VB1FitObj} taken in the argument.
}
\description{
access to the method in \code{OptMeth} field.
}
\author{
Cesaire J. K. Fouodo
}
\keyword{internal}
