% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{getSV.default}
\alias{getSV.default}
\title{Class \code{RegFitObj} (regression approach)}
\usage{
\method{getSV}{default}(rfo)
}
\arguments{
\item{rfo}{[\code{RegFitObj}(1)]\cr
object taken in the argument.}
}
\value{
\code{NULL}.
}
\description{
Accessor for the field \code{SV} for the object taken in an argument.
}
\author{
Cesaire J. K. Fouodo
}
\keyword{internal}

