% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{getSV.RegFitObj}
\alias{getSV.RegFitObj}
\title{\code{RegFitObj}}
\usage{
\method{getSV}{RegFitObj}(rfo)
}
\arguments{
\item{rfo}{[\code{RegFitObj}(1)]\cr
object taken in the argument.}
}
\value{
[\code{matrix}]
the matrix of support vectors.
}
\description{
Creator of the generic accessor \code{SV}.
}
\author{
Cesaire J. K. Fouodo
}
\keyword{internal}

