% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{getType.default}
\alias{getType.default}
\title{Default method of getType.}
\usage{
\method{getType}{default}(obj)
}
\arguments{
\item{obj}{[\code{Object(1)}]\cr
Object taken in argument.}
}
\value{
\code{NULL}.
}
\description{
Accessor for the field \code{Type} for the object taken in the argument.
}
\author{
Cesaire J. K. Fouodo
}
\keyword{internal}
