% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{getType.Kernel}
\alias{getType.Kernel}
\title{\code{Kernel}}
\usage{
\method{getType}{Kernel}(kern)
}
\arguments{
\item{kern}{[\code{\link{Kernel}(1)}]\cr
Object of class \code{Kernel}.}
}
\value{
Type of the kernel taken in the argument.
}
\description{
Accessor for the field \code{Type} of the object of class \code{Kernel} taken in an argument.
}
\author{
Cesaire J. K. Fouodo
}
\keyword{internal}
