% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{getType}
\alias{getType}
\title{Get the type of current object.}
\usage{
getType(obj)
}
\arguments{
\item{obj}{[\code{Object(1)}]\cr
Object taken in the argument.}
}
\description{
Creator of the generic accessor \code{getType}.
}
\author{
Cesaire J. K. Fouodo
}
\keyword{internal}
