% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vanbelle1.R
\name{getXtrain}
\alias{getXtrain}
\title{\code{VB1FitObj} (ranking approach)}
\usage{
getXtrain(vb1o)
}
\arguments{
\item{vb1o}{[\code{VB1FitObj}]\cr
Object taken in the argument.}
}
\description{
Creator of the generic accessor \code{getXtrain}.
}
\author{
Cesaire J. K. Fouodo
}
\keyword{internal}
