% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vanbelle1.R
\name{getXtrain.VB1FitObj}
\alias{getXtrain.VB1FitObj}
\title{\code{VB1FitObj} (ranking approach)}
\usage{
\method{getXtrain}{VB1FitObj}(vb1o)
}
\arguments{
\item{vb1o}{[\code{VB1FitObj}]\cr
Object taken in the argument.}
}
\value{
Xtrain [\code{matrix(1)}]
Field of the object of class \code{VB1FitObj} taken in the argument.
}
\description{
#' access to the train matrix of the \code{Xtrain} field.
}
\author{
Cesaire J. K. Fouodo
}
\keyword{internal}

