% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{Kernel}
\alias{Kernel}
\title{\code{Kernel}}
\usage{
Kernel(Type = NULL, Mat = NULL, Kernpar = NULL, bincat = NULL)
}
\arguments{
\item{Type}{[\code{character}]\cr
Type of kernel. Must be chosen from the following strings: \code{"lin_kernel"}, \code{"add_kernel"},
\code{"rbf_kernel"}, \code{"rbf4_kernel"} or \code{"poly_kernel"}.}

\item{Mat}{[\code{matrix(1)}]\cr
Kernel matrix.}

\item{Kernpar}{[\code{vector(1)}]\cr
Kernel parameters, when required.}

\item{bincat}{[\code{vector(1)}]\cr
Index of binary/categorical variables.}
}
\value{
[\code{\link{Kernel}(1)}]
Object of class Kernel, with elements:
\tabular{ll}{
   \code{Type} \tab type of kernel, \cr
   \code{Mat} \tab matrix used to perform differences between comparable data points. \cr
 }
}
\description{
Constructor of objects of class \code{Kernel}.
}
\author{
Cesaire J. K. Fouodo
}
\keyword{internal}

