% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{kernelMatrix}
\alias{kernelMatrix}
\title{Kernel}
\usage{
kernelMatrix(
  Xtrain,
  kernel_type = "lin_kernel",
  kernel_pars,
  Xt = NULL,
  bin_cat = integer(0)
)
}
\arguments{
\item{Xtrain}{[\code{matrix(1)}]\cr
Matrix of training data points.}

\item{kernel_type}{[\code{character(1)}]\cr
Type of kernel that is required.}

\item{kernel_pars}{[\code{vector(1)}]\cr
Parameters of kernels.}

\item{Xt}{[\code{matrix(1)}]\cr
Matrix of data points to be mapped.}

\item{bin_cat}{indexes of binary/categorical variables.}
}
\value{
[\code{\link{Kernel}(1)}]
Object of class Kernel, with elements:
\tabular{ll}{
   \code{Type} \tab type of kernel, \cr
   \code{Mat} \tab matrix used to compute differences between comparable data points. \cr
 }
}
\description{
computes the kernel matrix for elements taken in an argument.
}
\details{
\code{Kernel}
}
\author{
Cesaire J. K. Fouodo
}
\keyword{internal}
