% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diffmatrix.R
\name{makediff1}
\alias{makediff1}
\title{\code{Diffmatrix}}
\usage{
makediff1(Y, delta)
}
\arguments{
\item{Y}{[\code{vector(1)}]\cr
Ordered \code{vector} of survival times.}

\item{delta}{[\code{vector(1)}]\cr
Vector of status}
}
\value{
[\code{\link{Diffmatrix}(1)}]
Object of class \code{Diffmatrix} with elements:
 \tabular{ll}{
   \code{Type} \tab type of makediff function used to compute differences between neighbours. \cr
   \code{Mat} \tab matrix used to compute differences between comparable data points. \cr
 }
}
\description{
The data points are asssumed to be sorted by survival time. The comparison only takes place between two consecutivee observations when the first one is not
 censored (delta = 1).
}
\examples{
Y <- c(1,3,3.5,4,8); delta <- c(0,0,1,1,0); makediff1(Y, delta)
}
\seealso{
\code{\link{makediff2}} and \code{\link{makediff3}}
}

