% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diffmatrix.R
\name{makediff2}
\alias{makediff2}
\title{\code{Diffmatrix}}
\usage{
makediff2(Y, delta)
}
\arguments{
\item{Y}{[\code{vector(1)}]\cr
Ordered \code{vector} of survival times.}

\item{delta}{[\code{vector(1)}]\cr
Vector of status}
}
\value{
[\code{\link{Diffmatrix}(1)}]
Object of class \code{Diffmatrix} with elements:
 \tabular{ll}{
   \code{Type} \tab type of makediff function used to compute differences between neighbours. \cr
   \code{Mat} \tab matrix used to compute differences between comparable data points. \cr
 }
}
\description{
computes the matrix difference only between not censored data points (delta = 1).
 The data points are asssumed to be sorted by survival time and the difference
 is computed only if both the comparable data points are not censored.
}
\examples{
Y <- c(1,3,3.5,4,8); delta <- c(0,0,1,1,0); makediff2(Y, delta)
}
\seealso{
\code{\link{makediff1}} and \code{\link{makediff3}}
}
\keyword{internal}
