% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diffmatrix.R
\name{makediff3}
\alias{makediff3}
\title{\code{Diffmatrix}}
\usage{
makediff3(Y, delta)
}
\arguments{
\item{Y}{[\code{vector(1)}]\cr
Ordered \code{vector} of survival times.}

\item{delta}{[\code{vector(1)}]\cr
Vector of status.}
}
\value{
[\code{\link{Diffmatrix}(1)}]
Object of class \code{Diffmatrix} with elements:
 \tabular{ll}{
   \code{Type} \tab type of makediff function used to compute differences between neighbours. \cr
   \code{Mat} \tab matrix used to compute differences between comparable data points. \cr
 }
}
\description{
The first observation is assumed to be not censored (delta = 1).
 The Difference is computed between data point \code{i}
 and its neighbour that has the largest survival time but smaller than \code{y_i}, the survival time of \code{i}.
}
\examples{
Y <- c(1,3,3.5,4,8); delta <- c(0,0,1,1,0); makediff3(Y, delta)
}
\seealso{
\code{\link{makediff1}} and \code{\link{makediff2}}
}
