% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{predictHybrid}
\alias{predictHybrid}
\title{Survivalsvm predictions}
\usage{
predictHybrid(object, X_pred)
}
\arguments{
\item{object}{[\code{\link{survivalsvm}(1)}]\cr
survivalsvm object, fitted with \link{survivalsvm}.}

\item{X_pred}{[\code{matrix}]\cr
matrix of data points of interest.}
}
\value{
object of class survivalsvmprediction, with elements:
\tabular{ll}{
   \code{typeofsurvivalsvm} \tab type of the survivalsvm object that is fitted in model, \cr
   \code{typeofkernel} \tab type of kernel used to fit the model,\cr
   \code{parameterofkernel} \tab parameters of kernel used to fit the model,\cr
   \code{opt.meth} \tab program used to fit the model,\cr
   \code{predicted} \tab values predicted.\cr
 }
}
\description{
Maker of predictions based on model fitted using the hybrid approach of survival support vector machines.
}
\author{
Cesaire J. K. Fouodo
}
\keyword{internal}

