% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{predictRegFitObj}
\alias{predictRegFitObj}
\title{Survivalsvm predictions}
\usage{
predictRegFitObj(object, X_pred)
}
\arguments{
\item{object}{[\code{\link{survivalsvm}(1)}]\cr
makes predictions using a survivalsvm object fitted with \link{survivalsvm}.}

\item{X_pred}{[\code{matrix(1)}]\cr
matrix of data points of interest.}
}
\value{
[\code{survivalsvmprediction}(1)]
object of class survivalsvmprediction, with elements:
\tabular{ll}{
   \code{typeofsurvivalsvm} \tab type of the survivalsvm object that is fitted in model, \cr
   \code{typeofkernel} \tab type of kernel used to fit the model,\cr
   \code{parameterofkernel} \tab parameters of kernel that used to fit the model,\cr
   \code{opt.meth} \tab program used to fit the model,\cr
   \code{predicted} \tab values predicted.\cr
 }
}
\description{
Predictions based on model fitted using the regression approach of survival support vector machines.
}
\author{
Cesaire J. K. Fouodo
}
\keyword{internal}
