% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{predictVB2FitObj}
\alias{predictVB2FitObj}
\title{Survivalsvm predictions}
\usage{
predictVB2FitObj(object, X_pred)
}
\arguments{
\item{object}{[\code{\link{survivalsvm}(1)}]\cr
Object of class \code{survivalsvm}, fitted with \link{survivalsvm}.}

\item{X_pred}{[\code{matrix}]\cr
Matrix of data points used to make the prediction.}
}
\value{
object of class survivalsvmprediction, with elements:
\tabular{ll}{
   \code{typeofsurvivalsvm} \tab Type of the survivalsvm object that is fitted in model, \cr
   \code{typeofkernel} \tab type of kernel used to fit the model,\cr
   \code{parameterofkernel} \tab parameters of kernel that used to fit the model,\cr
   \code{opt.meth} \tab program used to fit the model,\cr
   \code{predicted} \tab values predicted.\cr
 }
}
\description{
Predictions based on model fitted using the ranking approach of survival support vector machines.
}
\keyword{internal}
