% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prints.R
\name{print.survivalsvm}
\alias{print.survivalsvm}
\title{print survivalsvm}
\usage{
\method{print}{survivalsvm}(x, ...)
}
\arguments{
\item{x}{[\code{survivalsvm(1)}]\cr
Object of class \code{survivalsvm} to be printed.}

\item{...}{[\code{any}]\cr
Further arguments passed to or from other methods.}
}
\description{
Prints  object of class \code{survivalsvm}.
}
\author{
Cesaire J. K. Fouodo
}
