% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prints.R
\name{print.survivalsvmprediction}
\alias{print.survivalsvmprediction}
\title{print survivalsvm}
\usage{
\method{print}{survivalsvmprediction}(x, ...)
}
\arguments{
\item{x}{[\code{survivalsvm(1)}]\cr
Object \code{survivalsvm}  to be printed.}

\item{...}{[\code{any}]\cr
Further arguments passed to or from other methods.}
}
\description{
Print objects of class \code{survivalsvm}.
}
\author{
Cesaire J. K. Fouodo
}
