% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prints.R
\name{printHybrid}
\alias{printHybrid}
\title{print survivalsvm}
\usage{
printHybrid(object, ...)
}
\arguments{
\item{object}{[\code{survivalsvm(1)}]\cr
Object \code{survivalsvm}  to be printed.}

\item{...}{[\code{any}]\cr
Further arguments passed to or from other methods.}
}
\description{
Print of object of class \code{Hybrid}. \code{Hybrid} is the class of models fitted using the hybrid approach of survival support vector machines.
}
\author{
Cesaire J. K. Fouodo
}
\keyword{internal}
