% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prints.R
\name{printRegFitObj}
\alias{printRegFitObj}
\title{print survivalsvm}
\usage{
printRegFitObj(object, ...)
}
\arguments{
\item{object}{[\code{survivalsvm(1)}]\cr
Object of class \code{survivalsvm} to be printed.}

\item{...}{[\code{any}]\cr
Further arguments passed to or from other methods.}
}
\description{
Print of object of class \code{RegFitObj}. \code{RegFitObj} is the class of models fitted using the regression approach of survival support vector machines.
}
\author{
Cesaire J. K. Fouodo
}
\keyword{internal}
