% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prints.R
\name{printVB2FitObj}
\alias{printVB2FitObj}
\title{print survivalsvm}
\usage{
printVB2FitObj(object, ...)
}
\arguments{
\item{object}{[\code{survivalsvm(1)}]\cr
Object \code{survivalsvm}  to be printed.}

\item{...}{[\code{any}]\cr
Further arguments passed to or from other methods.}
}
\description{
Print of object of class \code{RegFitObj}. \code{VB1FitObj} is the class of models fitted using the ranking approach of survival support vector machines.
}
\author{
Cesaire J. K. Fouodo
}
\keyword{internal}
