% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{RegFitObj}
\alias{RegFitObj}
\title{survivalsvm (regression approach)}
\usage{
RegFitObj(Beta = NULL, SV = NULL, Kernel = NULL, OptMeth = NULL,
  b0 = NULL)
}
\arguments{
\item{Beta}{[\code{vector(1)}]\cr
solution of the quadratic optimization problem of interest}

\item{SV}{[\code{matrix(1)}]\cr
support vector machines.}

\item{Kernel}{[\code{\link{Kernel}(1)}]\cr
object of class \code{Kernel}.}

\item{OptMeth}{[\code{character(1)}]\cr
program used to solve the optimization problem.}

\item{b0}{[\code{numeric(1)}]\cr
the estimated offset.}
}
\value{
[\code{RegFitObj}(1)]
object of class \code{RegFitObj} containing elements:
}
\description{
Constructs object of class \code{RegFitObj}.
}
\keyword{internal}

