% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vanbelle1.R
\name{setAlpha.default}
\alias{setAlpha.default}
\title{\code{VB1FitObj} (ranking approach)}
\usage{
\method{setAlpha}{default}(vb1o, alpha)
}
\arguments{
\item{vb1o}{[\code{VB1FitObj}]\cr
Object taken in the argument.}

\item{alpha}{[\code{vector(1)}]\cr
New offset.}
}
\value{
[\code{VB1FitObj}]
The object taken in the argument.
}
\description{
Default mutator of the field \code{Alpha} of the object taken in an argument.
}
\author{
Cesaire J. K. Fouodo
}
\keyword{internal}
