% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vanbelle1.R
\name{setAlpha}
\alias{setAlpha}
\title{\code{VB1FitObj} (ranking approach)}
\usage{
setAlpha(vb1o, alpha)
}
\arguments{
\item{vb1o}{[\code{VB1FitObj}]\cr
Object taken in the argument.}

\item{alpha}{[\code{vector(1)}]\cr
New value}
}
\description{
Creator of the generic mutator \code{setAlpha}.
}
\author{
Cesaire J. K. Fouodo
}
\keyword{internal}
