% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vanbelle1.R
\name{setAlpha.VB1FitObj}
\alias{setAlpha.VB1FitObj}
\title{\code{VB1FitObj} (ranking approach)}
\usage{
\method{setAlpha}{VB1FitObj}(vb1o, alpha)
}
\arguments{
\item{vb1o}{[\code{VB1FitObj}]\cr
Object of class \code{RegFitObj} taken in the argument.}

\item{alpha}{[\code{vector(1)}]\cr
Vector of solutions.}
}
\value{
modified version of the object taken in an argument.
}
\description{
Default mutator of the field \code{Alpha} of the object taken in an argument.
}
\keyword{internal}

