% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{setb0.default}
\alias{setb0.default}
\title{Class \code{RegFitObj} (regression approach)}
\usage{
\method{setb0}{default}(rfo, b0)
}
\arguments{
\item{rfo}{[\code{RegFitObj}(1)]\cr
object taken in the argument.}

\item{b0}{[\code{numeric(1)}]\cr
new offset value.}
}
\value{
[\code{RegFitObj}(1)]
the object taken in the argument.
}
\description{
Default mutator of the field \code{b0} of the object taken in an argument.
}
\author{
Cesaire J. K. Fouodo
}
\keyword{internal}

