% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{setb0}
\alias{setb0}
\title{Class \code{RegFitObj} (regression approach)}
\usage{
setb0(rfo, b0)
}
\arguments{
\item{rfo}{[\code{RegFitObj}(1)]\cr
object taken in the argument.}

\item{b0}{[\code{numeric(1)}]\cr
new value.}
}
\description{
Creator of the generic mutator \code{setb0}.
}
\author{
Cesaire J. K. Fouodo
}
\keyword{internal}
