% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{setb0.RegFitObj}
\alias{setb0.RegFitObj}
\title{\code{RegFitObj}}
\usage{
\method{setb0}{RegFitObj}(rfo, b0)
}
\arguments{
\item{rfo}{[\code{RegFitObj}(1)]\cr
object of class \code{RegFitObj} taken in the argument.}

\item{b0}{[\code{numeric(1)}]\cr
new offset.}
}
\value{
[\code{RegFitObj}(1)]
modified version of the object taken in the argument.
}
\description{
Default mutator of the field \code{b0} of the object taken in an argument.
}
\keyword{internal}
