% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{setBeta.default}
\alias{setBeta.default}
\title{\code{RegFitObj} (regression approach)}
\usage{
\method{setBeta}{default}(rfo, beta)
}
\arguments{
\item{rfo}{[\code{RegFitObj}(1)]\cr
object taken in the argument.}

\item{beta}{[\code{vector(1)}]\cr
new offset.}
}
\value{
the object taken in an argument.
}
\description{
Default mutator of the field \code{Beta} of the object taken in an argument.
}
\author{
Cesaire J. K. Fouodo
}
\keyword{internal}
