% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid.R
\name{setBeta.HybridObj}
\alias{setBeta.HybridObj}
\title{\code{Hybrid} (hybrid approach)}
\usage{
\method{setBeta}{HybridObj}(hybo, beta)
}
\arguments{
\item{hybo}{[\code{Hybrid(1)}]\cr
Object of class \code{Hybrid} taken in the argument.}

\item{beta}{[\code{vector(1)}]\cr
Index of binary/categorial variables.}
}
\value{
[\code{Hybrid(1)}]
Modified version of the object taken in the argument.
}
\description{
Default mutator of the field \code{Beta} of the object taken in an argument.
}
\keyword{internal}
