% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{setBeta}
\alias{setBeta}
\title{Class \code{RegFitObj} (regression approach)}
\usage{
setBeta(rfo, beta)
}
\arguments{
\item{rfo}{[\code{RegFitObj}(1)]\cr
object taken in the argument.}

\item{beta}{[\code{vector(1)}]\cr
new value.}
}
\description{
Creator of the generic mutator \code{setBeta}.
}
\author{
Cesaire J. K. Fouodo
}
\keyword{internal}
