% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{setBeta.RegFitObj}
\alias{setBeta.RegFitObj}
\title{\code{RegFitObj}}
\usage{
\method{setBeta}{RegFitObj}(rfo, beta)
}
\arguments{
\item{rfo}{[\code{RegFitObj}(1)]\cr
object of class \code{RegFitObj} taken in the argument.}

\item{beta}{[\code{vector(1)}]\cr
solutions of quadratic optimization problem.}
}
\value{
[\code{RegFitObj}(1)]
modified version of the object taken in the argument.
}
\description{
Default mutator of the field \code{Beta} of the object taken in an argument.
}
\keyword{internal}

