% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid.R
\name{setBetastar.default}
\alias{setBetastar.default}
\title{\code{Hybrid} (hybrid approach)}
\usage{
\method{setBetastar}{default}(hybo, betastar)
}
\arguments{
\item{hybo}{[\code{Hybrid(1)}]\cr
Object taken in the argument.}

\item{betastar}{[\code{vector(1)}]\cr
New value.}
}
\value{
[\code{Hybrid(1)}]
The object taken in the argument.
}
\description{
Default mutator of the field \code{Beta} of the object taken in an argument.
}
\author{
Cesaire J. K. Fouodo
}
\keyword{internal}
