% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid.R
\name{setBetastar.HybridObj}
\alias{setBetastar.HybridObj}
\title{\code{Hybrid} (hybrid approach)}
\usage{
\method{setBetastar}{HybridObj}(hybo, betastar)
}
\arguments{
\item{hybo}{[\code{Hybrid(1)}]\cr
Object of class \code{Hybrid} taken in the argument.}

\item{betastar}{[\code{vector(1)}]\cr
New value.}
}
\value{
[\code{Hybrid(1)}]
Modified version of the object taken in the argument.
}
\description{
Default mutator of the field \code{Betastar} of the object taken in an argument.
}
\keyword{internal}
