% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid.R
\name{setBetastar}
\alias{setBetastar}
\title{\code{Hybrid} (hybrid approach)}
\usage{
setBetastar(hybo, betastar)
}
\arguments{
\item{hybo}{[\code{Hybrid(1)}]\cr
Object taken in the argument.}

\item{betastar}{[\code{vector}]\cr
New value.}
}
\description{
Creator of the generic mutator \code{setBetastar}.
}
\author{
Cesaire J. K. Fouodo
}
\keyword{internal}

