% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{setBincat.default}
\alias{setBincat.default}
\title{\code{Kernel}}
\usage{
\method{setBincat}{default}(kern, bincat)
}
\arguments{
\item{kern}{[\code{\link{Kernel}(1)}]\cr
Object taken in the argument.}

\item{bincat}{[\code{vector(1)}]\cr
New index of binary/categorial variables.}
}
\value{
The object taken in the argument.
}
\description{
Default mutator of the field \code{bincat} of the object taken in an argument.
}
\author{
Cesaire J. K. Fouodo
}
\keyword{internal}
