% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{setBincat}
\alias{setBincat}
\title{\code{Kernel}}
\usage{
setBincat(kern, bincat)
}
\arguments{
\item{kern}{[\code{\link{Kernel}(1)}]\cr
Object of class \code{Kernel} taken in the argument.}

\item{bincat}{[\code{vector(1)}]\cr
Index of binary/categorial variables.}
}
\value{
The object taken in the argument.
}
\description{
Default mutator of the field \code{bincat} of the object taken in an argument.
}
\author{
Cesaire J. K. Fouodo
}
\keyword{internal}

