% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid.R
\name{setDelta.default}
\alias{setDelta.default}
\title{\code{Hybrid} (hybrid approach)}
\usage{
\method{setDelta}{default}(hybo, delta)
}
\arguments{
\item{hybo}{[\code{Hybrid(1)}]\cr
Object taken in the argument.}

\item{delta}{[\code{vector(1)}]\cr
New value}
}
\value{
[\code{Hybrid(1)}]
The object taken in the argument.
}
\description{
Default mutator of the field \code{delta} of the object taken in an argument.
}
\author{
Cesaire J. K. Fouodo
}
\keyword{internal}

