% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid.R
\name{setDelta.HybridObj}
\alias{setDelta.HybridObj}
\title{Creator of generic setor \code{setDelta}}
\usage{
\method{setDelta}{HybridObj}(hybo, delta)
}
\arguments{
\item{hybo}{[\code{Hybrid(1)}]\cr
Object of class \code{Hybrid} taken in the argument.}

\item{delta}{[\code{vector(1)}]\cr
New value.}
}
\value{
[\code{Hybrid(1)}]
Modified version of the object taken in the argument.
}
\description{
Creator of generic setor \code{setDelta}
}
\keyword{internal}
