% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid.R
\name{setDelta}
\alias{setDelta}
\title{\code{Hybrid} (hybrid approach)}
\usage{
setDelta(hybo, delta)
}
\arguments{
\item{hybo}{[\code{Hybrid(1)}]\cr
Object taken in the argument.}

\item{delta}{[\code{vector(1)}]\cr
New value.}
}
\description{
Creator of the generic mutator \code{setDelta}.
}
\author{
Cesaire J. K. Fouodo
}
\keyword{internal}

