% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vanbelle1.R
\name{setDifMat.default}
\alias{setDifMat.default}
\title{\code{VB1FitObj} (ranking approach)}
\usage{
\method{setDifMat}{default}(vb1o, dm)
}
\arguments{
\item{vb1o}{[\code{VB1FitObj}]\cr
Object taken in the argument.}

\item{dm}{[\code{\link{Diffmatrix}(1)}]\cr
New offset.}
}
\value{
[\code{VB1FitObj}]
The object taken in the argument.
}
\description{
Default mutator of the field \code{DifMat} of the object taken in an argument.
}
\author{
Cesaire J. K. Fouodo
}
\keyword{internal}
