% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vanbelle1.R
\name{setDifMat}
\alias{setDifMat}
\title{\code{VB1FitObj} (ranking approach)}
\usage{
setDifMat(vb1o, dm)
}
\arguments{
\item{vb1o}{[\code{VB1FitObj}]\cr
Object taken in the argument.}

\item{dm}{[\code{\link{Diffmatrix}(1)}]\cr
new value}
}
\description{
Creator of the generic mutator \code{setDifMat}.
}
\author{
Cesaire J. K. Fouodo
}
\keyword{internal}

