% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vanbelle1.R
\name{setDifMat.VB1FitObj}
\alias{setDifMat.VB1FitObj}
\title{\code{VB1FitObj} (ranking approach)}
\usage{
\method{setDifMat}{VB1FitObj}(vb1o, dm)
}
\arguments{
\item{vb1o}{[\code{VB1FitObj}]\cr
Object of class \code{RegFitObj} taken in the argument.}

\item{dm}{[\code{\link{Diffmatrix}(1)}]\cr
New value}
}
\value{
Modified version of the object taken in argument.
}
\description{
Default mutator of the field \code{DifMat} of the object taken in an argument.
}
\keyword{internal}

