% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{setKernel.default}
\alias{setKernel.default}
\title{\code{RegFitObj} (regression approach)}
\usage{
\method{setKernel}{default}(rfo, kernel)
}
\arguments{
\item{rfo}{[\code{RegFitObj}(1)]\cr
object taken in the argument.}

\item{kernel}{[\code{\link{Kernel}(1)}]\cr
new object of class \code{Kernel}.}
}
\value{
[\code{RegFitObj}(1)]
modified object.
}
\description{
setKernel.default
}
\keyword{internal}
