% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{setKernel}
\alias{setKernel}
\title{Class \code{RegFitObj} (regression approach)}
\usage{
setKernel(rfo, kernel)
}
\arguments{
\item{rfo}{[\code{RegFitObj}(1)]\cr
object taken in the argument.}

\item{kernel}{[\code{\link{Kernel}(1)}]\cr
new value.}
}
\description{
Creator of the generic mutator \code{setKernel}.
}
\author{
Cesaire J. K. Fouodo
}
\keyword{internal}
