% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{setKernel.RegFitObj}
\alias{setKernel.RegFitObj}
\title{\code{RegFitObj}}
\usage{
\method{setKernel}{RegFitObj}(rfo, kernel)
}
\arguments{
\item{rfo}{[\code{RegFitObj}(1)]\cr
object of class \code{RegFitObj} taken in the argument.}

\item{kernel}{[\code{\link{Kernel}(1)}]\cr
index of binary/categorial variables.}
}
\value{
[\code{RegFitObj}(1)]
modified version of the object taken in the argument.
}
\description{
Default mutator of the field \code{Kernel} of the object taken in an argument.
}
\keyword{internal}

