% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vanbelle1.R
\name{setKernel.VB1FitObj}
\alias{setKernel.VB1FitObj}
\title{\code{VB1FitObj} (ranking approach)}
\usage{
\method{setKernel}{VB1FitObj}(vb1o, kernel)
}
\arguments{
\item{vb1o}{[\code{VB1FitObj}]\cr
Object of class \code{RegFitObj} taken in the argument.}

\item{kernel}{[\code{\link{Diffmatrix}(1)}]\cr
New object of class \code{Kernel}.}
}
\value{
[\code{VB1FitObj}]
Modified version of the object taken in the argument.
}
\description{
Default mutator of the field \code{Kernel} of the object taken in an argument.
}
\keyword{internal}
