% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{setKernpar.default}
\alias{setKernpar.default}
\title{\code{Kernel}}
\usage{
\method{setKernpar}{default}(kern, kernpar)
}
\arguments{
\item{kern}{[\code{\link{Kernel}(1)}]\cr
Object taken in the argument.}

\item{kernpar}{[\code{vector(1)}]\cr
New kernel parameters.}
}
\value{
The object taken in the argument.
}
\description{
Default mutator of the field \code{Kernpar} of the object taken in an argument.
}
\author{
Cesaire J. K. Fouodo
}
\keyword{internal}
