% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{setKernpar.Kernel}
\alias{setKernpar.Kernel}
\title{\code{Kernel}}
\usage{
\method{setKernpar}{Kernel}(kern, kernpar)
}
\arguments{
\item{kern}{[\code{\link{Kernel}(1)}]\cr
Object of class \code{Kernel} taken in the argument.}

\item{kernpar}{[\code{vector(1)}]\cr
New kernel parameters.}
}
\value{
[\code{\link{Kernel}(1)}]
Object of class \code{Kernel} with elements:
 \tabular{ll}{
   \code{Type} \tab Type of kernel, \cr
   \code{Mat} \tab Kernel matrix, \cr
   \code{Kernpar} \tab Parameters of kernel, when required, \cr
   \code{bincat} \tab Index of binary/categorical variables, when required,
 }
}
\description{
Mutator of the field \code{Kernpar} of the object of class \code{Kernel} taken in an argument.
}
\author{
Cesaire J. K. Fouodo
}
\keyword{internal}
