% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{setKernpar}
\alias{setKernpar}
\title{\code{Kernel}}
\usage{
setKernpar(kern, kernpar)
}
\arguments{
\item{kern}{[\code{\link{Kernel}(1)}]\cr
Object of class \code{Kernel} taken in the argument.}

\item{kernpar}{[\code{vector(1)}]\cr
New kernel parameters.}
}
\description{
Default mutator of the field \code{Kernpar} of the object taken in an argument.
}
\author{
Cesaire J. K. Fouodo
}
\keyword{internal}

