% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{setMat}
\alias{setMat}
\title{\code{Kernel}}
\usage{
setMat(kern, mat)
}
\arguments{
\item{kern}{[\code{\link{Kernel}(1)}]\cr
Object taken in the argument.}

\item{mat}{[\code{matrix(1)}]\cr
New matrix.}
}
\description{
Creator of generic mutator \code{setMat}.
}
\author{
Cesaire J. K. Fouodo
}
\keyword{internal}

