% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{setOptMeth.default}
\alias{setOptMeth.default}
\title{Class \code{RegFitObj} (regression approach)}
\usage{
\method{setOptMeth}{default}(rfo, optmeth)
}
\arguments{
\item{rfo}{[\code{RegFitObj}(1)]\cr
object taken in the argument.}

\item{optmeth}{[\code{character(1)}]\cr
new name the of optimization program.}
}
\value{
[\code{RegFitObj}(1)]
the object taken in the argument.
}
\description{
Default mutator of the field \code{OptMeth} of the object taken in an argument.
}
\author{
Cesaire J. K. Fouodo
}
\keyword{internal}

