% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{setOptMeth}
\alias{setOptMeth}
\title{\code{RegFitObj} (regression approach)}
\usage{
setOptMeth(rfo, optmeth)
}
\arguments{
\item{rfo}{[\code{RegFitObj}(1)]\cr
object taken in the argument.}

\item{optmeth}{[\code{character(1)}]\cr
new value.}
}
\description{
Creator of the generic mutator \code{setOptMeth}.
}
\author{
Cesaire J. K. Fouodo
}
\keyword{internal}
