% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{setOptMeth.RegFitObj}
\alias{setOptMeth.RegFitObj}
\title{\code{RegFitObj}}
\usage{
\method{setOptMeth}{RegFitObj}(rfo, optmeth)
}
\arguments{
\item{rfo}{[\code{RegFitObj}(1)]\cr
object of class \code{RegFitObj} taken in the argument.}

\item{optmeth}{[\code{character(1)}]\cr
names the solver.}
}
\value{
[\code{RegFitObj}(1)]
modified version of the object taken in the argument.
}
\description{
Default mutator of the field \code{OptMeth} of the object taken in an argument.
}
\keyword{internal}

