% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vanbelle1.R
\name{setOptMeth.VB1FitObj}
\alias{setOptMeth.VB1FitObj}
\title{\code{VB1FitObj} (ranking approach)}
\usage{
\method{setOptMeth}{VB1FitObj}(vb1o, optmeth)
}
\arguments{
\item{vb1o}{[\code{VB1FitObj}]\cr
Object of class \code{RegFitObj} taken in the argument.}

\item{optmeth}{[\code{character(1)}]\cr
New value}
}
\value{
[\code{VB1FitObj}]
Modified version of the object taken in the argument.
}
\description{
Default mutator of the field \code{OptMeth} of the object taken in an argument.
}
\keyword{internal}
