% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{setSV.default}
\alias{setSV.default}
\title{\code{RegFitObj} (regression approach)}
\usage{
\method{setSV}{default}(rfo, sv)
}
\arguments{
\item{rfo}{[\code{RegFitObj}(1)]\cr
object taken in the argument.}

\item{sv}{[\code{matrix(1)}]\cr
new support vectors.}
}
\value{
[\code{RegFitObj}(1)]
the object taken in the argument.
}
\description{
Default mutator of the field \code{SV} of the object taken in an argument.
}
\author{
Cesaire J. K. Fouodo
}
\keyword{internal}

