% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{setSV}
\alias{setSV}
\title{Class \code{RegFitObj} (regression approach)}
\usage{
setSV(rfo, sv)
}
\arguments{
\item{rfo}{[\code{RegFitObj}(1)]\cr
object taken in the argument.}

\item{sv}{[\code{matrix(1)}]\cr
new value.}
}
\description{
Creator of the generic mutator \code{setSV}.
}
\author{
Cesaire J. K. Fouodo
}
\keyword{internal}
