% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{setSV.RegFitObj}
\alias{setSV.RegFitObj}
\title{\code{RegFitObj}}
\usage{
\method{setSV}{RegFitObj}(rfo, sv)
}
\arguments{
\item{rfo}{[\code{RegFitObj}(1)]\cr
object of class \code{RegFitObj} taken in the argument.}

\item{sv}{[\code{matrix(1)}]\cr
support vectors.}
}
\value{
[\code{RegFitObj}(1)]
modified version of the object taken in the argument.
}
\description{
Default mutator of the field \code{SV} of the object taken in an argument.
}
\keyword{internal}

