% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{setType.default}
\alias{setType.default}
\title{Class \code{Kernel}}
\usage{
\method{setType}{default}(obj, type)
}
\arguments{
\item{obj}{[\code{Object(1)}]\cr
Object taken in the argument.}

\item{type}{[\code{character(1)}]\cr
New type.}
}
\value{
Object taken in the argument.
}
\description{
Mutator of the field \code{Type} of the object taken in an argument.
}
\author{
Cesaire J. K. Fouodo
}
\keyword{internal}
