% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diffmatrix.R
\name{setType.Diffmatrix}
\alias{setType.Diffmatrix}
\title{\code{Diffmatrix}}
\usage{
\method{setType}{Diffmatrix}(obj, type)
}
\arguments{
\item{obj}{[\code{\link{Diffmatrix}(1)}]\cr
Object of class \code{Diffmatrix}.}

\item{type}{[\code{character(1)}]\cr
New type}
}
\value{
[\code{\link{Diffmatrix}(1)}]
Mutated object of class \code{Diffmatrix} containing elements.
}
\description{
Mutator for the type of objects of class \code{Diffmatrix}
}
\keyword{internal}
