% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{setType}
\alias{setType}
\title{Class \code{Kernel}}
\usage{
setType(obj, type)
}
\arguments{
\item{obj}{[\code{\link{Kernel}(1)}]\cr
Object taken in the argument.}

\item{type}{[\code{character(1)}]\cr
New type.}
}
\description{
Creator of the generic mutator \code{setType}.
}
\author{
Cesaire J. K. Fouodo
}
\keyword{internal}
