% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vanbelle1.R
\name{setXtrain.default}
\alias{setXtrain.default}
\title{\code{VB1FitObj} (ranking approach)}
\usage{
\method{setXtrain}{default}(vb1o, sv)
}
\arguments{
\item{vb1o}{[\code{VB1FitObj}]\cr
Object taken in the argument.}

\item{sv}{[\code{matrix(1)}]\cr
New value.}
}
\value{
[\code{VB1FitObj}]
The object taken in the argument.
}
\description{
Default mutator of the field \code{Xtrain} of the object taken in an argument.
}
\author{
Cesaire J. K. Fouodo
}
\keyword{internal}
