% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vanbelle1.R
\name{setXtrain}
\alias{setXtrain}
\title{\code{VB1FitObj} (ranking approach)}
\usage{
setXtrain(vb1o, sv)
}
\arguments{
\item{vb1o}{[\code{VB1FitObj}]\cr
Object taken in the argument.}

\item{sv}{[\code{matrix(1)}]\cr
new value}
}
\description{
Creator of the generic mutator \code{setXtrain}.
}
\author{
Cesaire J. K. Fouodo
}
\keyword{internal}
