% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vanbelle2.R
\name{VB2FitObj}
\alias{VB2FitObj}
\title{survivalsvm (ranking approach)}
\usage{
VB2FitObj(Alpha = NULL, Xtrain = NULL, DifMat = NULL, Kernel = NULL,
  OptMeth = NULL)
}
\arguments{
\item{Alpha}{[\code{vector(1)}]\cr
Solution of the quadratic optimization problem of interest.}

\item{Xtrain}{[\code{matrix(1)}]\cr
Matrix of training data points.}

\item{DifMat}{[\code{\link{Diffmatrix}(1)}]\cr
Matrix used to maked differences between neighbor points.}

\item{Kernel}{[\code{\link{Kernel}(1)}]\cr
Object of class \code{Kernel}.}

\item{OptMeth}{[\code{character(1)}]\cr
Program used to solve the optimization problem.}
}
\value{
\code{VB2FitObj}
Object of class \code{RegFitObj} containing elements:
\tabular{ll}{
   \code{Alpha} \tab solution of the quadratic optimization problem, \cr
   \code{Xtrain} \tab matrix of training points,\cr
   \code{DifMat} \tab matrix used to maked differences between neighbor points, \cr
   \code{Kernel} \tab kernel matrix, an object of class \code{Kernel},\cr
   \code{OptMeth} \tab program used to solve the quadratic optimization problem.\cr
 }
}
\description{
Constructs object of class \code{VB2FitObj}.
}
\keyword{internal}
