% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{getType.default}
\alias{getType.default}
\title{\code{Kernel}}
\usage{
\method{getType}{default}(kern)
}
\arguments{
\item{kern}{[\code{\link{Kernel}(1)}]\cr
Object taken in argument.}
}
\value{
\code{NULL}.
}
\description{
Accessor for the field \code{Type} for the object taken in the argument.
}
\author{
Cesaire J. K. Fouodo
}
\keyword{internal}
