% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diffmatrix.R
\name{getType.Diffmatrix}
\alias{getType.Diffmatrix}
\title{\code{Diffmatrix}.}
\usage{
\method{getType}{Diffmatrix}(dm)
}
\arguments{
\item{dm}{[\code{\link{Diffmatrix}(1)}]\cr
Object of class \code{Diffmatrix}.}
}
\value{
[\code{character(1)}]
Type of Diffmatrix
}
\description{
Mutator for objects of class \code{Diffmatrix}.
}
\keyword{internal}
