% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diffmatrix.R
\name{setMat.Diffmatrix}
\alias{setMat.Diffmatrix}
\title{\code{Diffmatrix}.}
\usage{
\method{setMat}{Diffmatrix}(dm, dmmat)
}
\arguments{
\item{dm}{[\code{\link{Diffmatrix}(1)}]\cr
Object of class \code{Diffmatrix}.}

\item{dmmat}{[\code{matrix(1)}]\cr
Matrix used to perform differences between comparable data points.}
}
\value{
[\code{\link{Diffmatrix}(1)}]
Mutated object of class \code{Diffmatrix} containing elements:
}
\description{
Mutator for the field \code{Mat} of objects of class \code{Diffmatrix}.
}
\author{
Cesaire J. K. Fouodo
}
\keyword{internal}
