% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{setMatrix.default}
\alias{setMatrix.default}
\title{\code{Kernel}}
\usage{
setMatrix.default(kern, kernmatrix)
}
\arguments{
\item{kern}{[\code{\link{Kernel}(1)}]\cr
Object taken in the argument.}

\item{kernmatrix}{[\code{matrix(1)}]\cr
New matrix.}
}
\value{
The object taken in the argument.
}
\description{
Mutator of the field \code{Mat} of the object taken in an argument.
}
\author{
Cesaire J. K. Fouodo
}
\keyword{internal}

