% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{setType.Kernel}
\alias{setType.Kernel}
\title{\code{Kernel}}
\usage{
\method{setType}{Kernel}(kern, kerntype)
}
\arguments{
\item{kern}{[\code{\link{Kernel}(1)}]\cr
Object taken in the argument.}

\item{kerntype}{[\code{character(1)}]\cr
Kerntype new type.}
}
\value{
Object of class \code{Kernel} with elements:
}
\description{
Mutator of the field \code{Type} of the object of class \code{Kernel} taken in an argument.
}
\author{
Cesaire J. K. Fouodo
}
\keyword{internal}
