% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivalsvm.R
\name{survivalsvm}
\alias{survivalsvm}
\title{survivalsvm}
\usage{
survivalsvm(formula = NULL, data = NULL, subset = NULL,
  type = "regression", diff.meth = NULL, gamma.mu = NULL,
  opt.meth = "quadprog", kernel = "lin_kernel", kernel.pars = NULL,
  time.variable.name = NULL, status.variable.name = NULL, sgf.sv = 5,
  sigf = 7, maxiter = 20, margin = 0.05, bound = 10, eig.tol = 1e-06,
  conv.tol = 1e-07, posd.tol = 1e-08)
}
\arguments{
\item{formula}{[\code{formula(1)}]\cr
Object of class \code{formula}. See \code{\link{formula}} for more details.}

\item{data}{[\code{data.frame(1)}]\cr
Object of class \code{data.frame} containing data points that will be used to fit the model.}

\item{subset}{[\code{vector(1)}]\cr
An index vector specifying the cases to be used in the training sample.}

\item{type}{[\code{character(1)}]\cr
String indicating which type of survival support vectors model is desired. This must be one
of the following strings: 'regression', 'vanbelle1', 'vanbelle2' or 'hybrid'.}

\item{diff.meth}{[\code{character(1)}]\cr
String indicating which of \code{'makediff1'}, \code{'makediff2'} or \code{'makediff3'}
is used in case of 'vanbelle1', 'vanbelle2' and 'hybrid'.}

\item{gamma.mu}{[\code{numeric(1)|vector(1)}]\cr
Parameters of regularization. Note that a vector with two parameters is required in case of \code{hybrid} approach. Just
one value is required in case of \code{regression}, \code{vanbelle1} or \code{vanbelle2}.}

\item{opt.meth}{[\code{character(1)}]\cr
Program used to solve the quadratic optimization problem. Either "\code{\link{quadprog}}" or "\code{\link{ipop}}".}

\item{kernel}{[\code{\link{Kernel}(1)}]\cr
Kernel used to fit the model: linear kern ('lin_kernel'), additive kernel ('add_kernel'),
radial basis kernels ('rbf_kernel') and the polynomial kernel ('poly_kernel').}

\item{kernel.pars}{[\code{vector(1)}]\cr
Parameters of kernel, when required.}

\item{time.variable.name}{[\code{character}]\cr
Name of the survival time variable in \code{data}, when given in argument.}

\item{status.variable.name}{[\code{character(1)}]\cr
Name of the status variable in \code{data}.}

\item{sgf.sv}{[\code{character(1)}]\cr
Number of decimal digits in the solution of the quadratic optimization problem.}

\item{sigf}{[\code{numeric(1)}]\cr
Used by \code{\link{ipop}}. See \code{\link{ipop}} for details.}

\item{maxiter}{[\code{integer(1)}]\cr
Used by \code{\link{ipop}}. See \code{\link{ipop}} for details.}

\item{margin}{[\code{numeric(1)}]\cr
Used by \code{\link{ipop}}. See \code{\link{ipop}} for details.}

\item{bound}{[\code{numeric(1)}]\cr
Used by \code{\link{ipop}}. See \code{\link{ipop}} for details.}

\item{eig.tol}{[\code{numeric(1)}]\cr
Used by \code{nearPD} for adjusting positive definiteness. See \code{\link{nearPD}} for detail.}

\item{conv.tol}{[\code{numeric(1)}]\cr
Used by \code{nearPD} for adjusting positive definiteness. See \code{\link{nearPD}} for detail.}

\item{posd.tol}{[\code{numeric(1)}]\cr
Used by \code{nearPD} for adjusting positive definiteness. See \code{\link{nearPD}} for detail.}
}
\value{
\code{survivalsvm}
Object of class \code{survivalsvm}, with elements:
\tabular{ll}{
   \code{call} \tab command calling this program, \cr
   \code{typeofsurvivalsvm} \tab type of survival support vector machines approach,\cr
   \code{model.fit} \tab the fitted survival model,\cr
   \code{var.names} \tab names of variables used.\cr
 }
}
\note{
This implementation is in part inspired by the \code{Matlab} toolbox \code{Survlab}
 (\href{http://user.it.uu.se/~kripe367/survlab/instruction.html}{\code{A Survival Analysis Toolbox}}).
}
\examples{

survivalsvm(Surv(time, status) ~ ., veteran, gamma.mu = 0.1)

survsvm.reg <- survivalsvm(formula = Surv(diagtime, status) ~ ., data = veteran,
                           type = "regression", gamma.mu = 0.1,
                           opt.meth = "ipop", kernel = "add_kernel")
                            
survsvm.vb2 <- survivalsvm(data = veteran, time.variable.name = "diagtime",
                           status.variable.name = "status", 
                           type = "vanbelle2", gamma.mu = 0.1,
                           opt.meth = "quadprog", diff.meth = "makediff3", 
                           kernel = "lin_kernel",
                           sgf.sv = 5, sigf = 7, maxiter = 20, 
                           margin = 0.05, bound = 10)
                            
}
\references{
\itemize{
 \item Van Belle, V., Pelcmans, K., Van Huffel S. and Suykens J. A.K. (2011a).
       Improved performance on high-dimensional survival data by application of Survival-SVM.
       Bioinformatics (Oxford, England) 27, 87-94.
 \item Van Belle, V., Pelcmans, K., Van Huffel S. and Suykens J. A.K. (2011b).
       Support vector methods for survival analysis: a comparaison between ranking and regression approaches.
       Artificial Intelligence in medecine 53, 107-118.
}
}
\seealso{
\link{predict.survivalsvm}
}
\author{
Cesaire J. K. Fouodo
}
